# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão de Colaboradores",
        "description": "Documentação da API de Gestão de Colaboradores utilizando Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/import-status/{id}": {
            "get": {
                "tags": [
                    "ImportStatus"
                ],
                "summary": "Obter o status de uma importação específica",
                "operationId": "b77dd26271c5ebd35661e06eeab4787c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do status de importação",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do status de importação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "completed"
                                        },
                                        "file_path": {
                                            "type": "string",
                                            "example": "storage/imports/file.csv"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Status de importação não encontrado ou acesso não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Status de importação não encontrado ou você não tem permissão para acessá-lo."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao buscar o status de importação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar status de importação."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token no formato 'Bearer {seu_token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}